Update tmux-sessionizer
Before it would not work if we tried to start a session with a similar
name as one we already have. For example if we would want to start
project-A when project-B already has a session, it would always open
project-B. By adding the equal sign we force `tmux has-session` to look
for an exact match in the open sessions.